requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair @ file:///home/conda/feedstock_root/build_artifacts/altair-split_1734244716962/work
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1756599151967/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1756198238354/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1756808288683/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1747811314515/work
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fonttools==4.59.2
frozenlist==1.7.0
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1735887193964/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1753444155689/work
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jsonschema_1755595646/work
jsonschema-specifications @ file:///tmp/tmpuvkyqc9y/src
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-text-splitters==0.3.11
langsmith==0.4.23
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
multidict==6.6.4
mypy_extensions==1.1.0
narwhals @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_narwhals_1756741653/work
nest-asyncio==1.6.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_numpy_1756343060/work/dist/numpy-2.3.2-cp312-cp312-linux_x86_64.whl#sha256=45c247eb27ec55f9e06d0544e3bdc8a9d711c021e9e34875c93f4aa7fcd1b771
orjson==3.11.3
packaging==25.0
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1755778906349/work
parso==0.8.4
pexpect==4.9.0
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1756853381948/work
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf @ file:///home/conda/feedstock_root/build_artifacts/protobuf_1751668317127/work/bazel-bin/python/dist/protobuf-6.31.1-cp312-abi3-linux_x86_64.whl#sha256=fd87fd790c588b35db1c2a03fbcac8381dc3bcf9372ca437dc29244489a16cc7
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck @ file:///home/conda/feedstock_root/build_artifacts/pydeck_1738346561094/work
Pygments==2.19.2
pyparsing==3.2.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
pyzmq==27.0.2
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1737836872/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
requests-toolbelt==1.0.0
rpds-py @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rpds-py_1756737801/work
setuptools==80.9.0
six==1.17.0
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1739781697784/work
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1756565351169/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1743640740204/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tornado==6.5.2
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
watchdog @ file:///home/conda/feedstock_root/build_artifacts/watchdog_1756135261048/work
wcwidth==0.2.13
wheel==0.45.1
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.24.0
